Fixed 'possible bug'
Most of the times req.user.save() was called it was marked 'with // possible bug?' due to the fact that it wasn't being used in an asynchronous context but should be an asynchronous function. However, I realised that if it is not used in combination with await, it will swallow errors and send an incorrect response to the client.